Fix update movie rating

+++ technologies/laravel/restful-api/database/migrations/2021_11_29_063022_create_movies_table.php
@@ -14,7 +14,7 @@ class CreateMoviesTable extends Migration {
         Schema::create('movies', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('rating')->nullable()->default(0);
+            $table->float('rating')->nullable()->default(0);
             $table->text("description")->nullable();
             $table->json("categories")->nullable();
             $table->json("actors")->nullable();
